Escape values of extra credentials and session properties
Correctly escape values of extra credentials and session properties when
building HTTP headers.

Change the format of extra credentials and session properties in the URL
to match the Trino JDBC and Python drivers - keys and values are
separated with a colon (`:`) and multiple entries are separated with a
semi-colon (`;`). Note: all special characters, including the semicolon,
must be URL escaped.